requirements.txt
beautifulsoup4==4.9.3
aiofiles==0.4.0
async-timeout==3.0.1
aiohttp==3.7.4
asyncio==3.4.3
asyncpg==0.22.0
dblpy==0.4.0
discord.py==1.6.0
tweepy==3.8.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
PyNaCl==1.3.0
pypapago==0.1.1.1
youtube-dl==2021.4.26
python-dotenv==0.17.0
filetype==1.0.7
xmltodict==0.12.0
ksoftapi==0.4.1
Pillow==8.2.0
datadog==0.41.0
pytz==2021.1
parsedatetime==2.6
Flask==1.1.2
scipy>=1.5.4
numexpr==2.7.3
